Removing duplicate test. Putting all the tests that have to do with the script in the same file.
