+++ tests/test_smart_dispatch.py
@@ -113,6 +113,45 @@ class TestSmartdispatcher(unittest.TestCase):
         assert_equal(exit_status_100, 2)
         assert_true(os.path.isdir(self.logs_dir))
 
+
+    def test_gpu_check(self):
+
+        argv = ['-x', '-g', '2', '-G', '1', '-C', '1', '-q', 'random', '-t', '00:00:10' ,'launch', 'echo', 'testing123']
+
+        # Test if the check fail
+        with self.assertRaises(SystemExit) as context:
+            smartdispatch_script.main(argv=argv)
+
+        self.assertTrue(context.exception.code, 2)
+
+        # Test if the test pass
+        argv[2] = '0'
+
+        try:
+            smartdispatch_script.main(argv=argv)
+        except SystemExit as e:
+            self.fail("The command failed the check, but it was supposed to pass.")
+
+
+    def test_cpu_check(self):
+
+        argv = ['-x', '-c', '2', '-C', '1', '-G', '1', '-t', '00:00:10', '-q', 'random', 'launch', 'echo', 'testing123']
+
+        # Test if the check fail
+        with self.assertRaises(SystemExit) as context:
+            smartdispatch_script.main(argv=argv)
+
+        self.assertTrue(context.exception.code, 2)
+
+        # Test if the test pass
+        argv[2] = '1'
+
+        try:
+            smartdispatch_script.main(argv=argv)
+        except SystemExit as e:
+            self.fail("The command failed the check, but it was supposed to pass.")
+
+
     @patch('subprocess.check_output')
     def test_launch_job_check(self, mock_check_output):
